requirements.txt
requests
pandas
textblob
duckdb
python-dotenv